Improve search algorithm
Track the best depth we've seen each state at to avoid traversing
unnecessary trees. This improvement make the example input resolve in a
reasonable time, but still takes ages for the real puzzle input. This is
because this is still doing a depth-first search, just pruning some
branches when we know they can't be fruitful.