Make setHash much more efficient.
CPU profiling revelaed that this was a hot point in the code, and that
using %#v with Printf is very expensive. Changing this to use a
bytes.buffer, and append the relevand data makes it much faster.
Additionally, reusing the same buffer each time saves on allocations,
and speeds it up further.

This change knocked c. 75% off the runtime for the part 1 input.