Improve memory efficiency.
Remove a bunch of allocations:

Use a sync.Pool for moves, and a global variable for the rates slice to
save allocating fresh ones every time.

Re-use the current and next lists instead of creating new ones each
time...

This improves things, but still isn't enough for the full input, which
still fails to return after a reasonable amount of time (although
memory growth is slowed.